added DefaultAvatar to @repo/ui and also small tweaks to table layout
